% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_design.R
\name{categorize_wells}
\alias{categorize_wells}
\title{Create well ranges for a design file}
\usage{
categorize_wells(start, end, nwells = 48, direction = "LR")
}
\arguments{
\item{start}{First well ID (e.g. "A1")}

\item{end}{Last well ID (e.g "B7")}

\item{nwells}{Total number of wells on the plate (currently only 48)}

\item{direction}{TB for top-to-bottom or LR for left-to-right}
}
\value{
string with a range of well IDs
}
\description{
Create well ranges for a design file
}
\examples{
# Wells from left to right, starting at A1 and ending at B4
categorize_wells("A1", "B4")
# Wells from top to bottom, starting at A1 and ending at B4
categorize_wells("A1", "B4", direction="TB")
}
