% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_electrode_dataset.R
\name{create_electrode_dataset}
\alias{create_electrode_dataset}
\title{Create Electrode Data set}
\usage{
create_electrode_dataset(data_path, recording_identifier)
}
\arguments{
\item{name}{data_path, recording_identifier}
}
\value{
Reformatted and filtered electrode burst list for use in analysis
}
\description{
Function reads and reformats electrode burst data obtained by the axis navigator tool for use in analysis
}
\examples{
test_data_1 <- create_electrode_dataset(data_path = "C:/Users/Emily Gordon/electrode_burst_list.csv", recording_identifier = "test_data_1")
}
