% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_synchrony_dataset.R
\name{create_synchrony_dataset}
\alias{create_synchrony_dataset}
\title{Create Synchrony Data Set}
\usage{
create_synchrony_dataset(data_path, heatmap_condition)
}
\arguments{
\item{name}{data_path, heatmap_condition}
}
\value{
Reformatted and filtered neural metric dataset for use in synchrony analysis
}
\description{
Function reads in neural metric output dataset and reformats for use in synchrony index analysis
}
\examples{
baseline_data <- create_synchrony_dataset(data_path = "C:/Users/Emily Gordon/baseline_neuralMetrics.csv", heatmap_condition = "Baseline")
}
