% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrode_mean_burst_ISI.R
\name{electrode_mean_burst_ISI}
\alias{electrode_mean_burst_ISI}
\title{Electrode Mean Burst ISI}
\usage{
electrode_mean_burst_ISI(
  data = output_table,
  time_lower_bound = 0,
  time_upper_bound = Inf,
  recording_identifier
)
}
\arguments{
\item{name}{data, time_lower_bound, time_upper_bound, recording_identifier}
}
\value{
electrode burst list table with three additional columns (mean, SD and SEM mean ISI within a burst (sec) for single electrodes within a given time interval and recording)
}
\description{
Function reads in electrode burst list data and creates new columns with the mean, SD, and SEM mean ISI within a burst (sec) recorded by single electrodes within a given time interval and recording
}
\examples{
output_table <- electrode_mean_burst_ISI(data = output_table, 0, 30, recording_identifier = "test_data_1")
}
