% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_outliers.R
\name{find_outliers}
\alias{find_outliers}
\title{Find and label outliers}
\usage{
find_outliers(
  df,
  col_parameter = "Parameter",
  col_measurement = "Measurement_normalized",
  col_concentration = "Concentration"
)
}
\arguments{
\item{df}{Input data frame}

\item{col_parameter}{Name of the column (as a string) that contains parameters}

\item{col_measurement}{Name of the column (as a string) that contains (normalized) measurements}

\item{col_concentration}{Name of the column (as a string) that contains concentration values}
}
\description{
Detect outliers with the adjusted box plot function
that takes skewed data into account. Label outliers
in a separate column and return the dataframe.
}
