% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_heatmap_data.R
\docType{data}
\name{input_heatmap_data}
\alias{input_heatmap_data}
\title{Input dataset for MEA_heatmap function}
\format{
A data frame with 24 rows and 3 variables
\describe{
  \item{Well}{}
  \item{synchrony_index: Baseline}{}
  \item{synchrony_index: Agonist Challenge}{}
}
}
\source{
<https://www.github.com/egordon2/MEAanalysis>
}
\usage{
input_heatmap_data
}
\description{
A data set with well average synchrony index data for a baseline and comparison experimental condition.
}
\keyword{datasets}
