% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_MEA_data.R
\docType{data}
\name{input_MEA_data}
\alias{input_MEA_data}
\title{Input dataset for electrode and well MEA parameter functions.}
\format{
A data frame with 8886 rows and 36variables:
\describe{
  \item{Electrode}{}
  \item{Time_s}{}
  \item{Size_spikes}{}
  \item{Duration_s}{}
  \item{Recording_identifier}{}
  \item{well}{}
}
}
\source{
<https://www.github.com/egordon2/MEAanalysis>
}
\usage{
input_MEA_data
}
\description{
A data set containing electrode burst list data filtered and reformated for use in analysis.
}
\keyword{datasets}
