% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_designfile.R
\name{parse_designfile}
\alias{parse_designfile}
\title{Parse designfile}
\usage{
parse_designfile(path)
}
\arguments{
\item{path}{the location of the design file on the path}
}
\value{
list with two objects: metadata, containing information on the experiment
and design, containing a dataframe with cells and corresponding groups.
}
\description{
Parses a text file containing the design of the
}
