% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_MEA_file.R
\name{parse_MEA_file}
\alias{parse_MEA_file}
\title{Parse an MEA output file}
\usage{
parse_MEA_file(path)
}
\arguments{
\item{path}{Path to the MEA output file}
}
\value{
list with [1] header, [2] well averages, and [3] electrode data
}
\description{
From the MEA output file, generate data frames with header info,
well average data, and electrode data.
}
