% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_MEA_data.R
\name{treatment_ratio}
\alias{treatment_ratio}
\title{Calculate treatment ratios}
\usage{
treatment_ratio(exposurepath, baselinepath)
}
\arguments{
\item{exposurepath}{Path to file with exposure data}

\item{baselinepath}{Path to file with baseline data}
}
\description{
Calculate the treatment ratio between exposure and baseline
of MEA output files.
(Currently incomplete)
}
