% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_stats.R
\name{treatment_stats}
\alias{treatment_stats}
\title{Generate statistics for a treatment dataframe}
\usage{
treatment_stats(ratio_df)
}
\arguments{
\item{ratio_df}{resulting dataframe of the `treatment_ratio` function}
}
\value{
dataframe with summary statistics
}
\description{
Generate statistics for a treatment dataframe
}
