% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/well_barchart.R
\name{well_barchart}
\alias{well_barchart}
\title{Well Bar chart}
\usage{
well_barchart(
  data = output_table,
  well_parameter,
  well_filter = "A1|A2|A3|A4|A5|A6|B1|B2|B3|B4|B5|B6|C1|C2|C3|C4|C5|C6|D1|D2|D3|D4|D5|D6",
  statistic = se
)
}
\arguments{
\item{name}{data, well_parameter, well_filter, statistic}
}
\value{
bar chart of calculated MEA parameters by single electrode for a given time interval and recording
}
\description{
Function creates a bar chart of calculated MEA parameters by well for a given time interval and recording
}
\examples{
well_barchart(data = output_table, well_parameter = "burst_duration", well_filter = "A1|A2|A3|A4", statistic = se)
well_barchart(data = output_table, well_parameter = "spikes_per_burst", well_filter = "A1|A2|A3|A4", statistic = se)
well_barchart(data = output_table, well_parameter = "mean_burst_ISI", well_filter = "A1|A2|A3|A4", statistic = se)
well_barchart(data = output_table, well_parameter = "number_of_bursts", well_filter = "A1|A2|A3|A4", statistic = se)
}
