% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/well_burst_duration.R
\name{well_burst_duration}
\alias{well_burst_duration}
\title{Well Burst Duration}
\usage{
well_burst_duration(
  data = output_table,
  time_lower_bound = 0,
  time_upper_bound = Inf,
  recording_identifier
)
}
\arguments{
\item{name}{data, time_lower_bound, time_upper_bound, recording_identifier}
}
\value{
electrode burst list table with three additional columns (mean, SD and SEM burst duration recorded for each well within a given time interval of a recording)
}
\description{
Function reads in electrode burst list data and creates new columns with the mean, SD, and SEM burst duration recorded for each well within a given time interval of a recording
}
\examples{
output_table <- well_burst_duration(data = output_table, 0, 30, recording_identifier = "test_data_1")
}
