% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/well_number_of_bursts.R
\name{well_number_of_bursts}
\alias{well_number_of_bursts}
\title{Well Number of Bursts}
\usage{
well_number_of_bursts(
  data = output_table,
  time_lower_bound = 0,
  time_upper_bound = Inf,
  recording_identifier
)
}
\arguments{
\item{name}{data, time_lower_bound, time_upper_bound, recording_identifier}
}
\value{
electrode burst list table with three additional columns (mean, SD and SEM number of bursts recorded for each well within a given time interval of a recording)
}
\description{
Function reads in electrode burst list data and creates new columns with the mean, SD, and SEM number of bursts recorded for each well within a given time interval of a recording
}
\examples{
output_table <- well_number_of_bursts(data = output_table, 0, 30, recording_identifier = "test_data_1")
}
