% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_electrode_dataset.R
\name{create_electrode_dataset}
\alias{create_electrode_dataset}
\title{Create Electrode Data set}
\usage{
create_electrode_dataset(data_path, recording_identifier)
}
\arguments{
\item{data_path}{Include path to electrode burst dataset csv file produced by the axis navigator tool in quotation marks.}

\item{recording_identifier}{Include a unique identifier for this specific electrode burst dataset which will be added to the 'Recording_identifier' column. Other functions will use this column to filter for a specific recording.}
}
\value{
A reformatted and filtered electrode burst list for use in analysis.
}
\description{
Function reads in and reformats the electrode burst list dataset produced by the axis navigator tool for use in analysis.
 The data will be filtered to only contain information relating to single electrodes and burst characteristics.
 A 'Recording_identifier' column is added to enable filtering for a specific recording when used alongside other functions.
 A column to identify the well a burst was recorded within will also be added.
}
\examples{
output_electrode_burst <- create_electrode_dataset(
                                    data_path = system.file("extdata", "input_electrode_burst.csv",
                                                            package = "MEAanalysis"),
                                    recording_identifier = "burst_recording_1")
}
