% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_design.R
\name{exp_design}
\alias{exp_design}
\title{Generate experimental design file}
\usage{
exp_design(date, expID, dir = ".")
}
\arguments{
\item{date}{date (string, preferably formatted as YYYYMMDD)}

\item{expID}{identifier for the experiment (string)}

\item{dir}{directory where the file should be saved}
}
\description{
With information provided by the user, a file is created
that formalizes the experimental design so it can be used to
analyse the MEA file.
}
