% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_electrode_barchart.R
\name{single_electrode_barchart}
\alias{single_electrode_barchart}
\title{Single Electrode Bar chart}
\usage{
single_electrode_barchart(
  data = output_table,
  electrode_parameter,
  electrode_filter =
    "A1_|A2_|A3_|A4_|A5_|A6_|B1_|B2_|B3_|B4_|B5_|B6_|C1_|C2_|C3_|C4_|C5_|C6_|D1_|D2_|D3_|D4_|D5_|D6_",
  statistic = se
)
}
\arguments{
\item{name}{data, electrode_parameter, electrode_filter, statistic}
}
\value{
bar chart of calculated MEA parameters by single electrode for a given time interval and recording
}
\description{
Function creates a bar chart of calculated MEA parameters by single electrode for a given time interval and recording
}
\examples{
single_electrode_barchart(data = output_table, electrode_parameter = "burst_duration", electrode_filter = "A1_", statistic = se)
single_electrode_barchart(data = output_table, electrode_parameter = "spikes_per_burst", electrode_filter = "A1_", statistic = se)
single_electrode_barchart(data = output_table, electrode_parameter = "mean_burst_ISI", electrode_filter = "A1_", statistic = se)
single_electrode_barchart(data = output_table, electrode_parameter = "number_of_bursts", electrode_filter = "A1_")
}
