% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_MEA_data.R, R/treatment_ratio.R
\name{treatment_ratio}
\alias{treatment_ratio}
\title{Calculate treatment ratios}
\usage{
treatment_ratio(exposurepath, baselinepath, treatment)

treatment_ratio(exposurepath, baselinepath, treatment)
}
\arguments{
\item{exposurepath}{Path to file with exposure data}

\item{baselinepath}{Path to file with baseline data}

\item{treatment}{Path to file with treatment lay-out}
}
\value{
file with [1] treatment ratio per parameter per well
}
\description{
Calculate the treatment ratio between exposure and baseline
of MEA output files.
(Currently incomplete)
}
