% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_ratio.R
\name{treatment_ratio}
\alias{treatment_ratio}
\title{Calculate the treatment ratio between exposure and baseline of MEA output files}
\usage{
treatment_ratio(exposurepath, baselinepath, designpath, save = F, path = ".")
}
\arguments{
\item{exposurepath}{Path to file with exposure data}

\item{baselinepath}{Path to file with baseline data}

\item{designpath}{Path to file with treatment lay-out}

\item{save}{Boolean to indicate whether result should be saved}

\item{path}{Path where file should be saved}
}
\value{
data frame with treatment ratio per parameter per well (only if `save` is false)
}
\description{
This function calculates the difference between exposure and baseline MEA output files
and adds to this the information from a design file.
The output can either be saved as a csv file, or it is returned by the function.
}
